fix: moved method to list private files from folder with subfolders
