+++ server/src/main/java/com/epam/aidial/core/server/AiDial.java
@@ -25,7 +25,6 @@ import com.epam.aidial.core.server.service.VertxTimerService;
 import com.epam.aidial.core.server.service.codeinterpreter.CodeInterpreterService;
 import com.epam.aidial.core.server.service.schemarichapps.ApplicationResourcesReplacementService;
 import com.epam.aidial.core.server.service.schemarichapps.PublicationEnrichmentService;
-import com.epam.aidial.core.server.service.schemarichapps.ResourceListingService;
 import com.epam.aidial.core.server.token.TokenStatsTracker;
 import com.epam.aidial.core.server.tracing.DialTracingFactory;
 import com.epam.aidial.core.server.upstream.UpstreamRouteProvider;
@@ -140,9 +139,8 @@ public class AiDial {
             NotificationService notificationService = new NotificationService(resourceService, encryptionService);
             ResourceOperationService resourceOperationService = new ResourceOperationService(applicationService,
                     resourceService, invitationService, shareService, lockService);
-            ResourceListingService resourceListingService = new ResourceListingService(encryptionService, resourceService);
             PublicationEnrichmentService publicationEnrichmentService = new PublicationEnrichmentService(applicationService, encryptionService,
-                    resourceService, resourceListingService, configStore);
+                    resourceService, configStore);
             ApplicationResourcesReplacementService
                     applicationResourcesReplacementService =
                     new ApplicationResourcesReplacementService(configStore, encryptionService, resourceService);
